:hammer: refactor area picker and add to list
