+++ src/modules/List.tsx
@@ -8,9 +8,11 @@ import {
     TableRowColumn
 } from "material-ui/Table";
 import { Person } from "../interfaces/apiInterfaces";
+import SelectArea from "../components/selectArea";
 
 interface TableProps {
     area: number | null;
+    handler: any;
 }
 /**
  * A simple table demonstrating the hierarchy of the `Table` component and its sub-components.
@@ -20,41 +22,51 @@ const TableExampleSimple = (props: TableProps) => {
         // fetch people
         const people: Person[] = [];
         return (
-            <Table>
-                <TableHeader displaySelectAll={false} adjustForCheckbox={false}>
-                    <TableRow>
-                        <TableHeaderColumn>Name</TableHeaderColumn>
-                        <TableHeaderColumn>Address</TableHeaderColumn>
-                        <TableHeaderColumn>District</TableHeaderColumn>
-                        <TableHeaderColumn>Status</TableHeaderColumn>
-                        <TableHeaderColumn>Extra Info</TableHeaderColumn>
-                    </TableRow>
-                </TableHeader>
-                <TableBody displayRowCheckbox={false} stripedRows={true}>
-                    <TableRow>
-                        <TableRowColumn>John Smith</TableRowColumn>
-                        <TableRowColumn>123 Tester Lane</TableRowColumn>
-                        <TableRowColumn>Watermelon</TableRowColumn>
-                        <TableRowColumn>Safe</TableRowColumn>
-                        <TableRowColumn>
-                            What is life? Baby don't hurt me.
-                        </TableRowColumn>
-                    </TableRow>
-                    {people.map(person => {
-                        return (
-                            <Row
-                                name={person.name}
-                                status={person.safe}
-                                address={person.district}
-                                extraInfo={person.extra_info}
-                            />
-                        );
-                    })}
-                </TableBody>
-            </Table>
+            <div style={{ margin: "1em 2em" }}>
+                <SelectArea handler={props.handler} />
+                <Table>
+                    <TableHeader
+                        displaySelectAll={false}
+                        adjustForCheckbox={false}
+                    >
+                        <TableRow>
+                            <TableHeaderColumn>Name</TableHeaderColumn>
+                            <TableHeaderColumn>Address</TableHeaderColumn>
+                            <TableHeaderColumn>District</TableHeaderColumn>
+                            <TableHeaderColumn>Status</TableHeaderColumn>
+                            <TableHeaderColumn>Extra Info</TableHeaderColumn>
+                        </TableRow>
+                    </TableHeader>
+                    <TableBody displayRowCheckbox={false} stripedRows={true}>
+                        <TableRow>
+                            <TableRowColumn>John Smith</TableRowColumn>
+                            <TableRowColumn>123 Tester Lane</TableRowColumn>
+                            <TableRowColumn>Watermelon</TableRowColumn>
+                            <TableRowColumn>Safe</TableRowColumn>
+                            <TableRowColumn>
+                                What is life? Baby don't hurt me.
+                            </TableRowColumn>
+                        </TableRow>
+                        {people.map(person => {
+                            return (
+                                <Row
+                                    name={person.name}
+                                    status={person.safe}
+                                    address={person.district}
+                                    extraInfo={person.extra_info}
+                                />
+                            );
+                        })}
+                    </TableBody>
+                </Table>
+            </div>
         );
     } else {
-        return <div>Select An Area in Home first!</div>;
+        return (
+            <div style={{ margin: "1em 2em" }}>
+                <SelectArea handler={props.handler} />
+            </div>
+        );
     }
 };
